Reinitialisation of the master branch
As our previous work was based on an outdated design I decide to reinitialise our repository.
Included are a few documents I created over the weekend using the Selection page version 6 design.

As of this point the initial markup has been completed, although more work needs to be done to styling, interactivity and instantiation of table elements.