Created ClassManagement.css and modified other pages (3 weeks of work)
I created a Class Management page. It use files in the AJAX folder to update the   displayed class asycronously.

I've also created import and export PHP scripts in Downloads and DataBase, and changed the markup and styling on pages like the TeacherHomePage.php, SortSelections.php and ModulesAndSpins.php.
in particular ModulesAndSpins.php had a lot of changes so I created a copy called ModulesAndSpins.1.php with appropriately named css and js files. I made the curriculum tally bar sticky meaning I had to overhaul a large part of the markup, styling and scripts. The page is not quite done; I still need to get the deselect buttons working again, but everything else, to my knowlegde, is working.